requirements.txt
Flask==2.3.2
flask-cors==4.0.0
google-cloud-storage==2.16.0
gunicorn==21.2.0
